Moved endpoints to seperate folder cleaning program
